Rename the GitHub Release fetcher
