+++ lib/getReleases.js
